Refactor extension and add stale download lock recovery
- Steal .download.lock files older than 5 minutes after crashes
- Extract hooks.ts (debounce, toast lifecycle) and preview.tsx
- Add httpsGetJson, buildSearchArgs, mergedFindrEnv helpers in utils.ts
- Slim search.tsx to orchestration only (~450 lines)
