+++ extensions/findr/src/search.tsx
@@ -11,21 +11,9 @@ import {
   open,
   showInFinder,
 } from "@raycast/api";
-import { useState, useMemo, useEffect, useRef } from "react";
-import {
-  existsSync,
-  mkdirSync,
-  rmSync,
-  readdirSync,
-  openSync,
-  readSync,
-  closeSync,
-  renameSync,
-} from "fs";
-import { createHash } from "crypto";
-import { execFile, ChildProcess } from "child_process";
-import { tmpdir } from "os";
-import { join } from "path";
+import { useState, useMemo, useEffect } from "react";
+import { existsSync } from "fs";
+import { execFile } from "child_process";
 import { SearchResponse, SearchResult } from "./types";
 import {
   getFindrPath,
@@ -33,13 +21,15 @@ import {
   getMaxResults,
   getFindrEnv,
   getCustomPaths,
-  formatFileSize,
-  formatRelativeDate,
   getFileIcon,
   trackInteraction,
   parseSearchStdout,
+  mergedFindrEnv,
+  buildSearchArgs,
 } from "./utils";
 import { openBugReport } from "./bug-report";
+import { useDebouncedValue, useFailureToast, useAnimatedToast } from "./hooks";
+import { ResultDetail } from "./preview";
 
 function getResultIcon(result: SearchResult): string {
   if (result.is_dir) return "📁";
@@ -131,17 +121,19 @@ function ResultActions({ result }: { result: SearchResult }) {
   );
 }
 
-/** Debounce hook — prevents useExec from firing on every keystroke */
-function useDebouncedValue(value: string, delayMs: number): string {
-  const [debounced, setDebounced] = useState(value);
-  const timerRef = useRef<ReturnType<typeof setTimeout>>(undefined);
-
-  useEffect(() => {
-    timerRef.current = setTimeout(() => setDebounced(value), delayMs);
-    return () => clearTimeout(timerRef.current);
-  }, [value, delayMs]);
-
-  return debounced;
+function ResultItem({ result }: { result: SearchResult }) {
+  return (
+    <List.Item
+      icon={getResultIcon(result)}
+      title={result.filename}
+      accessories={result.is_dir ? [{ tag: "Folder" }] : []}
+      quickLook={
+        result.is_dir ? undefined : { path: result.path, name: result.filename }
+      }
+      detail={<ResultDetail result={result} />}
+      actions={<ResultActions result={result} />}
+    />
+  );
 }
 
 export default function SearchFiles() {
@@ -149,9 +141,9 @@ export default function SearchFiles() {
   const [findrPath, setFindrPath] = useState(() => getFindrPath());
   const maxResults = getMaxResults();
   const findrEnv = getFindrEnv();
+  const execEnv = useMemo(() => mergedFindrEnv(findrEnv), [findrEnv]);
   const binaryExists = useMemo(() => existsSync(findrPath), [findrPath]);
 
-  // Auto-download findr binary from GitHub Releases if not present
   useEffect(() => {
     if (binaryExists) return;
     let cancelled = false;
@@ -178,18 +170,13 @@ export default function SearchFiles() {
     };
   }, [binaryExists]);
 
-  // Debounce search input — prevents racing subprocesses on fast typing
   const debouncedQuery = useDebouncedValue(query, 300);
   const isSearchReady = debouncedQuery.trim().length >= 2;
 
-  // Phase 1: Fast search (no semantic) — fires immediately on debounced query
   const [searchData, setSearchData] = useState<SearchResponse | null>(null);
   const [searchLoading, setSearchLoading] = useState(false);
   const [searchError, setSearchError] = useState<Error | null>(null);
-  const errorToastRef = useRef<Toast | null>(null);
-  const indexingToastRef = useRef<Toast | null>(null);
 
-  // Drop stale search errors as soon as the user edits the query.
   useEffect(() => {
     setSearchError(null);
   }, [query]);
@@ -207,16 +194,8 @@ export default function SearchFiles() {
     let killed = false;
     const child = execFile(
       findrPath,
-      [
-        "search",
-        debouncedQuery,
-        "--json",
-        "--limit",
-        String(maxResults),
-        "--no-semantic",
-        "--no-sync",
-      ],
-      { env: { ...process.env, ...findrEnv } },
+      buildSearchArgs(debouncedQuery, maxResults, { sync: false }),
+      { env: execEnv },
       (err, stdout) => {
         if (killed) return;
         if (err) {
@@ -240,39 +219,38 @@ export default function SearchFiles() {
       killed = true;
       child.kill();
     };
-  }, [debouncedQuery, isSearchReady, binaryExists]);
+  }, [
+    debouncedQuery,
+    isSearchReady,
+    binaryExists,
+    findrPath,
+    maxResults,
+    execEnv,
+  ]);
 
-  // Phase 2: Semantic search — fires 1s after user stops typing, merges results
-  const debouncedSemantic = useDebouncedValue(query, 1300); // 300ms base + 1000ms extra
+  const debouncedSemantic = useDebouncedValue(query, 1300);
   const isSemanticReady = debouncedSemantic.trim().length >= 2;
 
   useEffect(() => {
-    if (!isSemanticReady || !binaryExists) return;
-    // Only fire if fast results already loaded (avoid double-loading on first render)
-    if (!searchData) return;
+    if (!isSemanticReady || !binaryExists || !searchData) return;
 
     let killed = false;
     const child = execFile(
       findrPath,
-      [
-        "search",
-        debouncedSemantic,
-        "--json",
-        "--limit",
-        String(maxResults),
-        "--no-sync",
-      ],
-      { env: { ...process.env, ...findrEnv }, timeout: 4000 },
+      buildSearchArgs(debouncedSemantic, maxResults, {
+        semantic: true,
+        sync: false,
+      }),
+      { env: execEnv, timeout: 4000 },
       (err, stdout) => {
-        if (killed) return;
-        if (err) return; // silently skip — fast results already showing
+        if (killed || err) return;
         try {
           const semanticData = parseSearchStdout(stdout);
           if (semanticData.total_results > 0) {
             setSearchData(semanticData);
           }
         } catch {
-          // ignore — fast results stay
+          /* fast results stay */
         }
       },
     );
@@ -281,11 +259,16 @@ export default function SearchFiles() {
       killed = true;
       child.kill();
     };
-  }, [debouncedSemantic, isSemanticReady, binaryExists, searchData !== null]);
+  }, [
+    debouncedSemantic,
+    isSemanticReady,
+    binaryExists,
+    searchData !== null,
+    findrPath,
+    maxResults,
+    execEnv,
+  ]);
 
-  // Recent files: stale-then-refresh pattern
-  // 1. Instant: show cached results (--no-sync, <50ms)
-  // 2. Background: sync + fresh results, replace if different
   const [recentData, setRecentData] = useState<SearchResponse | null>(null);
   const [recentLoading, setRecentLoading] = useState(true);
 
@@ -309,27 +292,26 @@ export default function SearchFiles() {
         if (clearLoading) setRecentLoading(false);
       };
 
-    const execEnv = { env: { ...process.env, ...findrEnv } };
-    const runRecentSearch = (noSync: boolean, clearLoading: boolean) => {
-      const args = ["search", "", "--json", "--limit", String(maxResults)];
-      if (noSync) args.push("--no-sync");
-      return execFile(findrPath, args, execEnv, onRecentResults(clearLoading));
-    };
-
-    // Phase 1: instant cached results (no sync)
-    const staleChild = runRecentSearch(true, true);
-    // Phase 2: background sync + fresh results (replaces stale)
-    const freshChild = runRecentSearch(false, false);
+    const staleChild = execFile(
+      findrPath,
+      buildSearchArgs("", maxResults, { sync: false }),
+      { env: execEnv },
+      onRecentResults(true),
+    );
+    const freshChild = execFile(
+      findrPath,
+      buildSearchArgs("", maxResults),
+      { env: execEnv },
+      onRecentResults(false),
+    );
 
     return () => {
       cancelled = true;
       staleChild.kill();
       freshChild.kill();
     };
-  }, [findrPath, binaryExists]);
+  }, [findrPath, binaryExists, maxResults, execEnv]);
 
-  // Auto-index new custom paths: detect preference changes, debounce 10s, index only new paths.
-  // cancelled flag prevents the async .then() callback from spawning work after unmount.
   const customPaths = getCustomPaths();
   useEffect(() => {
     if (!binaryExists || !customPaths) return;
@@ -351,7 +333,6 @@ export default function SearchFiles() {
       const newPaths = currentList.filter((p) => !storedSet.has(p));
 
       if (newPaths.length > 0) {
-        // Debounce 10s to let user finish editing
         timer = setTimeout(() => {
           if (cancelled) return;
           let completed = 0;
@@ -359,7 +340,6 @@ export default function SearchFiles() {
             execFile(findrPath, ["index", "add-path", p], (err) => {
               if (cancelled) return;
               if (!err) completed++;
-              // Only mark as indexed after all subprocesses finish successfully
               if (completed === newPaths.length) {
                 LocalStorage.setItem(
                   "findr_custom_paths",
@@ -370,7 +350,6 @@ export default function SearchFiles() {
           }
         }, 10000);
       } else if (currentList.length > 0 && !stored) {
-        // First time: store current paths without triggering index
         LocalStorage.setItem("findr_custom_paths", currentList.join(","));
       }
     });
@@ -379,68 +358,24 @@ export default function SearchFiles() {
       cancelled = true;
       if (timer) clearTimeout(timer);
     };
-  }, [customPaths, binaryExists]);
+  }, [customPaths, binaryExists, findrPath]);
 
   const isTyping = query.length > 0;
   const showRecent = !isTyping && recentData?.mode === "recent";
   const recentResults = recentData?.results || [];
-
   const searchResults = searchData?.results || [];
   const elapsed = searchData?.elapsed_ms ?? 0;
   const isIndexing = searchData?.mode === "indexing";
-
-  // Show loading only during active search, not during initial debounce
   const isLoading = isTyping
     ? searchLoading || (query !== debouncedQuery && isSearchReady)
     : recentLoading;
 
-  useEffect(() => {
-    if (!searchError) return;
-
-    errorToastRef.current?.hide();
-    showToast({
-      style: Toast.Style.Failure,
-      title: "Search failed",
-      message: searchError.message,
-      primaryAction: {
-        title: "Copy Error",
-        shortcut: { modifiers: ["cmd"], key: "t" },
-        onAction: async (toast) => {
-          await Clipboard.copy(searchError.message);
-          toast.hide();
-        },
-      },
-    }).then((toast) => {
-      errorToastRef.current = toast;
-    });
-
-    return () => {
-      errorToastRef.current?.hide();
-      errorToastRef.current = null;
-    };
-  }, [searchError]);
-
-  useEffect(() => {
-    if (!isIndexing) {
-      indexingToastRef.current?.hide();
-      indexingToastRef.current = null;
-      return;
-    }
-
-    indexingToastRef.current?.hide();
-    showToast({
-      style: Toast.Style.Animated,
-      title: "Building index for the first time...",
-      message: "This takes ~25 seconds. Search again shortly.",
-    }).then((toast) => {
-      indexingToastRef.current = toast;
-    });
-
-    return () => {
-      indexingToastRef.current?.hide();
-      indexingToastRef.current = null;
-    };
-  }, [isIndexing]);
+  useFailureToast(searchError);
+  useAnimatedToast(
+    isIndexing,
+    "Building index for the first time...",
+    "This takes ~25 seconds. Search again shortly.",
+  );
 
   if (!binaryExists) {
     return (
@@ -530,270 +465,5 @@ export default function SearchFiles() {
         </List.Section>
       )}
     </List>
-  );
-}
-
-function ResultItem({ result }: { result: SearchResult }) {
-  return (
-    <List.Item
-      icon={getResultIcon(result)}
-      title={result.filename}
-      accessories={result.is_dir ? [{ tag: "Folder" }] : []}
-      quickLook={
-        result.is_dir ? undefined : { path: result.path, name: result.filename }
-      }
-      detail={<ResultDetail result={result} />}
-      actions={<ResultActions result={result} />}
-    />
-  );
-}
-
-const IMAGE_TYPES = new Set([
-  "png",
-  "jpg",
-  "jpeg",
-  "gif",
-  "webp",
-  "svg",
-  "heic",
-  "tiff",
-  "bmp",
-  "ico",
-]);
-const TEXT_PREVIEW_TYPES = new Set([
-  "md",
-  "txt",
-  "csv",
-  "html",
-  "xml",
-  "json",
-  "yaml",
-  "yml",
-  "toml",
-  "ini",
-  "cfg",
-  "conf",
-  "log",
-  "sql",
-  "sh",
-  "zsh",
-  "bash",
-  "rs",
-  "ts",
-  "tsx",
-  "js",
-  "jsx",
-  "py",
-  "go",
-  "rb",
-  "java",
-  "c",
-  "cpp",
-  "h",
-  "css",
-  "scss",
-]);
-const MAX_PREVIEW_BYTES = 8192;
-const MAX_PREVIEW_LINES = 40;
-
-const RENDER_AS_PLAIN = new Set(["txt", "md"]);
-
-function readTextPreview(path: string, ext: string): string {
-  try {
-    const buf = Buffer.alloc(MAX_PREVIEW_BYTES);
-    const fd = openSync(path, "r");
-    const bytesRead = readSync(fd, buf, 0, MAX_PREVIEW_BYTES, 0);
-    closeSync(fd);
-    const raw = buf.slice(0, bytesRead).toString("utf-8");
-    // Cut at last complete line to avoid mid-line truncation
-    const lastNewline = raw.lastIndexOf("\n");
-    const text = lastNewline > 0 ? raw.slice(0, lastNewline) : raw;
-    const allLines = text.split("\n");
-    const truncated =
-      allLines.length > MAX_PREVIEW_LINES || bytesRead >= MAX_PREVIEW_BYTES;
-    const lines = allLines.slice(0, MAX_PREVIEW_LINES);
-    const content = lines.join("\n");
-
-    const truncNote = truncated ? "\n\n---" : "";
-
-    // Plain text / markdown: render as-is (no code box)
-    if (RENDER_AS_PLAIN.has(ext)) {
-      return content + truncNote;
-    }
-    // CSV: render as markdown table (first 5 columns, 20 rows)
-    if (ext === "csv") {
-      const rows = lines.slice(0, 20);
-      if (rows.length >= 2) {
-        const parsed = rows.map((r) =>
-          r.split(",").map((c) => c.trim().replace(/^"|"$/g, "")),
-        );
-        const maxCols = Math.min(parsed[0].length, 5);
-        const trim = (s: string) => (s.length > 20 ? s.slice(0, 18) + ".." : s);
-        const header =
-          "| " + parsed[0].slice(0, maxCols).map(trim).join(" | ") + " |";
-        const sep = "| " + Array(maxCols).fill("---").join(" | ") + " |";
-        const body = parsed
-          .slice(1)
-          .map((r) => "| " + r.slice(0, maxCols).map(trim).join(" | ") + " |")
-          .join("\n");
-        const extra =
-          parsed[0].length > 5
-            ? `\n\n*+${parsed[0].length - 5} more columns*`
-            : "";
-        return header + "\n" + sep + "\n" + body + extra;
-      }
-    }
-    // Everything else: syntax-highlighted code block
-    const lang =
-      ext === "py"
-        ? "python"
-        : ext === "rs"
-          ? "rust"
-          : ext === "js" || ext === "jsx"
-            ? "javascript"
-            : ext === "ts" || ext === "tsx"
-              ? "typescript"
-              : ext === "rb"
-                ? "ruby"
-                : ext === "yml" || ext === "yaml"
-                  ? "yaml"
-                  : ext;
-    return "```" + lang + "\n" + content + "\n```" + truncNote;
-  } catch {
-    return "";
-  }
-}
-
-function ResultDetail({ result }: { result: SearchResult }) {
-  const [preview, setPreview] = useState<string>("");
-
-  // Async file preview — avoids blocking the render thread with synchronous I/O.
-  // All branches fall through to a single cleanup that cancels pending work.
-  useEffect(() => {
-    let cancelled = false;
-    let timer: ReturnType<typeof setTimeout> | undefined;
-    let qlChild: ChildProcess | undefined;
-
-    setPreview("");
-
-    // Image: inline markdown (no I/O needed)
-    if (result.file_type && IMAGE_TYPES.has(result.file_type)) {
-      setPreview(
-        `![preview](file://${result.path.split("/").map(encodeURIComponent).join("/")}?raycast-height=250)`,
-      );
-    }
-    // PDF thumbnail via qlmanage (async)
-    else if (result.file_type === "pdf") {
-      const hash = createHash("md5").update(result.path).digest("hex");
-      const thumbDir = join(tmpdir(), "findr-thumbs");
-      const thumbPath = join(thumbDir, `${hash}.png`);
-
-      if (existsSync(thumbPath)) {
-        setPreview(`![preview](file://${thumbPath})\n\n`);
-      } else {
-        const qlOutDir = join(thumbDir, hash);
-        mkdirSync(qlOutDir, { recursive: true });
-        if (!cancelled) {
-          qlChild = execFile(
-            "qlmanage",
-            ["-t", result.path, "-s", "600", "-o", qlOutDir],
-            { timeout: 3000 },
-            () => {
-              if (cancelled) return;
-              try {
-                const pngFile = readdirSync(qlOutDir).find((name) =>
-                  name.endsWith(".png"),
-                );
-                if (pngFile) {
-                  renameSync(join(qlOutDir, pngFile), thumbPath);
-                }
-              } catch {
-                /* qlmanage may have failed */
-              }
-              try {
-                rmSync(qlOutDir, { recursive: true });
-              } catch {
-                /* best effort */
-              }
-              if (existsSync(thumbPath)) {
-                setPreview(`![preview](file://${thumbPath})\n\n`);
-              }
-            },
-          );
-        }
-      }
-    }
-    // Text/code preview (async read via next tick)
-    else if (
-      result.file_type &&
-      TEXT_PREVIEW_TYPES.has(result.file_type) &&
-      !result.is_dir
-    ) {
-      timer = setTimeout(() => {
-        if (cancelled) return;
-        const text = readTextPreview(result.path, result.file_type!);
-        if (text && !cancelled) {
-          setPreview(text + "\n\n");
-        }
-      }, 0);
-    }
-
-    return () => {
-      cancelled = true;
-      if (timer) clearTimeout(timer);
-      if (qlChild) qlChild.kill();
-    };
-  }, [result.path]);
-
-  let markdown = "";
-
-  // Content snippet first (most important when searching)
-  if (result.content_snippet) {
-    const sanitized = result.content_snippet
-      .replace(/[\\`*_{}[\]()#+\-.!|<>~]/g, "\\$&")
-      .replace(/\n/g, "\n> ");
-    markdown += `> ${sanitized}\n\n`;
-  }
-
-  markdown += preview;
-
-  return (
-    <List.Item.Detail
-      markdown={markdown || undefined}
-      metadata={
-        <List.Item.Detail.Metadata>
-          <List.Item.Detail.Metadata.Label title="Path" text={result.path} />
-          <List.Item.Detail.Metadata.Separator />
-          {result.is_dir ? (
-            <List.Item.Detail.Metadata.Label title="Type" text="FOLDER" />
-          ) : (
-            result.file_type && (
-              <List.Item.Detail.Metadata.Label
-                title="Type"
-                text={result.file_type.toUpperCase()}
-              />
-            )
-          )}
-          {result.size_bytes && !result.is_dir && (
-            <List.Item.Detail.Metadata.Label
-              title="Size"
-              text={formatFileSize(result.size_bytes)}
-            />
-          )}
-          {result.modified && (
-            <List.Item.Detail.Metadata.Label
-              title="Modified"
-              text={formatRelativeDate(result.modified)}
-            />
-          )}
-          {result.interactions > 0 && (
-            <List.Item.Detail.Metadata.Label
-              title="Interactions"
-              text={String(result.interactions)}
-            />
-          )}
-        </List.Item.Detail.Metadata>
-      }
-    />
   );
 }
